% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_functions.R
\name{rbiosnp_readGVCF}
\alias{rbiosnp_readGVCF}
\title{rbiosnp_readGVCF}
\usage{
rbiosnp_readGVCF(file, verbose = TRUE)
}
\arguments{
\item{file}{Input \code{vcf} file.}

\item{verbose}{Verbose.}
}
\value{
A \code{gvcf} object containing the information from the VCF file.
}
\description{
The function loads a \code{vcf} file resulted from the GATK suite.
}
\details{
This function deals with the \code{GVCF} primarily generated by the HaplotypeCaller command.
         GVCF stands for Genomic VCF. A GVCF is a kind of VCF;
         so the basic format specification is the same as for a regular VCF (see the spec documentation here),
         but a Genomic VCF contains extra information.
         The key difference between a regular VCF and a GVCF is that the GVCF has records for all sites,
         whether there is a variant call there or not.
}
\examples{
\dontrun{
cohort_gvcf <- rbiosnp_readGVCF(file = "rawSNP.g.vcf")
}
}
